only first connected player can change settings and start game

+++ Assets/Scripts/Settings/GameSettingTime.cs
@@ -6,13 +6,13 @@
 
     protected override string Get()
     {
-        return ((long)field.GetValue(game.settings) / 1000000000).ToString() + "s";
+        return ((long)field.GetValue(game.settings) / 1000000000.0).ToString() + "s";
     }
 
     protected override void Set(string text)
     {
         object settings = game.settings;
-        field.SetValue(settings, long.Parse(text.Replace("s", "")) * 1000000000);
+        field.SetValue(settings, (long)(double.Parse(text.Replace("s", "")) * 1000000000.0));
         game.settings = (GameSettings)settings;
     }
 }
